Implement recipe discovery
This implements recipe discovery, which is a recursive process to
discover all relevant package recipes for a given project repository.

This process starts at the initial repo, adding each of the recipes
advertised there (in a toplevel `recipes.eld` file). Then, it
traverses upstream to the repositories that each of these recipes
points to, repeating the process until either (a) no advertised
recipes are found, or (b) the repository has already been visited
during traversal.

Recipe discovery allows a project repository to be the single source
of truth on how to (1) obtain and (2) build the packages it provides,
including any necessary dependencies that may be hosted elsewhere. Any
recipes for dependencies that are not discovered through this process
are presumed to be available on standard package archives such as
ELPA.