Distinguish primary and pointer recipes
A project repo is only responsible for the correctness and
completeness of its own recipes, that is, the ones that describe how
to build the packages contained in the repo itself. For dependencies,
the project is only responsible for providing sufficient information
to "point" to the correct source repositories for the dependencies.
These could be complete and valid recipes but they need not be, as
long as the addressing portion (e.g., host and repo) is valid.

We now distinguish these primary and pointer recipes, adding the
former and traversing the latter (ignoring any non-addressing fields).
When these pointed-to repos are traversed, their primary recipes will
be discovered there and then added.

This also adds an `:auto` recipe field to distinguish canonical
recipes that are to be automatically managed by Elacarte from bespoke
recipes that are added directly by the user (either manually or via
`elacarte-add-recipe`). These recipes added by the user will not
generally be overwritten unless an explicit `replace` flag is set.