seamlessly keep recipes up to date
The package manager (i.e., Straight) is responsible for determining
what to build and which version of it to build. But before it actually
builds the package, we want to ensure that the recipe that will be
used to build it is up to date with what's advertised by the version
of the repo about to be built.

For this purpose, we simply use Straight's provided pre-build hook to
add the current recipe to the user's `recipes.eld`. For recipes that
are managed automatically by Elacarte, the new recipe overwrites the
older one in the user's recipes.

This enables us to seamlessly keep recipes updated, deferring to the
source repository itself as the source of truth on how its packages
should be built, without relying on third parties such as centralized
package repositories for this essential information.