refactor to make recipe discovery file- instead of list-oriented
Formerly, the core recursive logic was housed in a function that
traversed a list of recipes. Since we don't expect this to happen
outside of a traversal of an actual recipes file, we now house this
core logic in a function addressing this explicit scenario.

This also abstracts the "clean room" installation of packages that's
necessary to determine attributes of the installation such as the
unique local repo name, making it available as a separate utility, and
also making it more "functional" by defining a specific output for
this clean room installation, viz., the normalized recipe.