Serve recipes directly from recipes.eld
Formerly, building a recipe repository involved creating a directory
containing one file per recipe (similar to MELPA). This makes reading
recipes more efficient as we only need to read the file containing the
specific recipe we seek. But this approach necessitates an extra step
whenever recipes are updated, to rebuild the recipe repository, in
order for these changes to take effect.

As Elacarte's recipes.eld isn't meant to hold every conceivable recipe
but specifically those for packages that are installed in the user's
configuration, we do not anticipate this file being extraordinarily
large, so that the efficiency of reads isn't a concern, and serving
recipes directly from recipes.eld eliminates the intermediate step of
rebuilding the recipe repository, so that package recipe changes take
effect immediately when the recipes.eld file is updated.

This also renames the local recipe repository from "XELPA" to
"Elacarte Cookbook."