Add a low-level utility to just add all recipes in a file
This is useful in tools like elci which take on the role of the user
in being the ultimate authority on what recipes to use. They may
choose to define their recipes in various files, and these are to be
used verbatim.

Ultimately, Elacarte gives users ultimate authority in defining
recipes, and just provides various supporting conveniences such as
automatically managing recipes and keeping them up to date,
discovering relevant recipes given a project repository of interest,
and so on.

This new utility enables bypassing these supporting conveniences and
directly specifying the recipes in a file --- a convenient convention
for the purposes of automation tools.