Support specifying a primary recipe in a dependent repo
This should generally be avoided, as the dependent repo cannot be
responsible for maintaining an accurate primary recipe for a
dependency. But having the ability to do this when needed is still
valuable, e.g., at least during development, but also when the
dependency repo cannot or will not advertise its recipes.